feat(decorators): Refactored swagger path mapping
