+++ src/utils/route-handler.util.ts
@@ -1,9 +1,18 @@
-import { DECORATOR_METADATA_ENUM } from "../constants/decorator.constants";
+import { AppConfig } from "../app-config";
+import {
+  DECORATOR_METADATA_ENUM,
+  DTO_DECORATOR_METADATA_ENUM,
+  ROUTE_DECORATOR_METADATA_ENUM,
+} from "../constants/decorator.constants";
 import {
   ClassType,
   DeepReadonly,
+  EnumForType,
   IRouteArgMetadata,
+  IRouteBody,
   IRouterHandlerArgs,
+  ISwaggerReferenceSchema,
+  ISwaggerSchema,
 } from "../interfaces";
 
 /**
@@ -81,3 +90,218 @@ export const generateRouteHandlerArgs = (
   }
   return routeArgs;
 };
+
+/**
+ * Process dependency for a DTO class and add it to swagger schema
+ * @param this - this keyword referenced to AppConfig class instance
+ * @param target - DTO class, whose dependency to add in swagger schema
+ * @returns void
+ */
+// eslint-disable-next-line func-style
+function processDependenciesOfSchema(
+  // eslint-disable-next-line @typescript-eslint/prefer-readonly-parameter-types
+  this: AppConfig,
+  target: any,
+): void {
+  const dependencies: { type: unknown; isEnum: boolean }[] | undefined =
+    Reflect.getMetadata(
+      DTO_DECORATOR_METADATA_ENUM.DTO_SCHEMA_DEPENDENCY,
+      target.prototype,
+    );
+
+  if (!dependencies) {
+    return;
+  }
+
+  while (dependencies.length > 0) {
+    const dependency = dependencies.shift() as { type: any; isEnum: boolean };
+    const nestedDependencies =
+      Reflect.getMetadata(
+        DTO_DECORATOR_METADATA_ENUM.DTO_SCHEMA_DEPENDENCY,
+        dependency.type.prototype,
+      ) ?? [];
+    dependencies.push(...nestedDependencies);
+
+    // eslint-disable-next-line no-invalid-this
+    const clonedSwaggerConfig = this.swaggerConfigCopy;
+
+    let dtoSchema: ISwaggerSchema = {
+      type: "string",
+      enum: Object.values(dependency),
+    };
+
+    if (!dependency.isEnum) {
+      dtoSchema = Reflect.getMetadata(
+        DTO_DECORATOR_METADATA_ENUM.DTO_SCHEMA,
+        dependency.type.prototype,
+      ) ?? { type: "object" };
+    }
+
+    clonedSwaggerConfig.components.schemas = {
+      ...(clonedSwaggerConfig.components.schemas ?? {}),
+      [dependency.type.name]: dtoSchema,
+    };
+    // eslint-disable-next-line no-invalid-this
+    this.swaggerConfigCopy = clonedSwaggerConfig;
+  }
+}
+
+/**
+ * This function is used to map schema from parameter type or method return type to swagger
+ * @param this - this keyword referenced to AppConfig class instance
+ * @param type - Type associated with method or parameter
+ * @returns swagger schema type and swagger reference schema string
+ */
+/* eslint-disable no-invalid-this */
+// eslint-disable-next-line func-style
+export function getReferenceSchema(
+  // eslint-disable-next-line @typescript-eslint/prefer-readonly-parameter-types
+  this: AppConfig,
+  type: any,
+): ISwaggerReferenceSchema | ISwaggerSchema {
+  let schema: ISwaggerReferenceSchema | ISwaggerSchema | undefined = undefined;
+  const clonedSwaggerConfig = this.swaggerConfigCopy;
+
+  if (
+    type.name === "Object" ||
+    type.name === "Boolean" ||
+    type.name === "Number" ||
+    type.name === "String" ||
+    type.name === "Function"
+  ) {
+    schema = {
+      $ref: `#/components/schemas/${type.name}`,
+    };
+  } else if (type.name === "undefined") {
+    schema = {
+      $ref: "#/components/schemas/Undefined",
+    };
+  }
+
+  if (schema !== undefined) {
+    return schema;
+  }
+
+  const dtoSchema: ISwaggerSchema = Reflect.getMetadata(
+    DTO_DECORATOR_METADATA_ENUM.DTO_SCHEMA,
+    type.prototype,
+  ) ?? { type: "object" };
+  const dtoClassName = type.name;
+
+  schema = {
+    $ref: `#/components/schemas/${dtoClassName}`,
+  };
+
+  clonedSwaggerConfig.components.schemas = {
+    ...(clonedSwaggerConfig.components.schemas ?? {}),
+    [dtoClassName]: dtoSchema,
+  };
+  this.swaggerConfigCopy = clonedSwaggerConfig;
+
+  processDependenciesOfSchema.call(this, type);
+  return schema;
+}
+/* eslint-enable no-invalid-this */
+
+/**
+ * Process specified route's request or response body
+ * @param this - this keyword referenced to AppConfig class instance
+ * @param bodyType - Route body type. To determine if it is request or response body
+ * @param routeHandlerDetails - Route handler details
+ * @returns Either swagger schema or undefined if no route body is specified explicitly
+ */
+/* eslint-disable no-invalid-this */
+// eslint-disable-next-line func-style
+export function processRouteBody(
+  // eslint-disable-next-line @typescript-eslint/prefer-readonly-parameter-types
+  this: AppConfig,
+  bodyType: ROUTE_DECORATOR_METADATA_ENUM,
+  routeHandlerDetails: Readonly<{
+    controller: ClassType;
+    routeHandlerName: string;
+  }>,
+): ISwaggerReferenceSchema | ISwaggerSchema | undefined {
+  const routeBodyMetadata: IRouteBody | undefined = Reflect.getMetadata(
+    bodyType,
+    routeHandlerDetails.controller.prototype,
+    routeHandlerDetails.routeHandlerName,
+  );
+  if (routeBodyMetadata === undefined) {
+    return undefined;
+  }
+
+  let swaggerSchema: ISwaggerReferenceSchema | ISwaggerSchema | undefined;
+  if (typeof routeBodyMetadata.type === typeof EnumForType) {
+    swaggerSchema = {
+      type: "string",
+      deprecated: routeBodyMetadata.options.deprecated,
+      nullable: routeBodyMetadata.options.nullable,
+      required: routeBodyMetadata.options.required,
+    };
+
+    if (routeBodyMetadata.options.isArray === true) {
+      swaggerSchema.type = "array";
+      swaggerSchema.items = {
+        type: "string",
+        enum: Object.values(routeBodyMetadata.type),
+      };
+    } else {
+      swaggerSchema.enum = Object.values(routeBodyMetadata.type);
+    }
+  } else if (
+    routeBodyMetadata.type === "Boolean" ||
+    routeBodyMetadata.type === "Number" ||
+    routeBodyMetadata.type === "Object" ||
+    routeBodyMetadata.type === "String" ||
+    routeBodyMetadata.type === "Undefined"
+  ) {
+    swaggerSchema = {
+      // eslint-disable-next-line @typescript-eslint/no-base-to-string
+      $ref: `#/components/schemas/${routeBodyMetadata.type}`,
+      deprecated: routeBodyMetadata.options.deprecated,
+      nullable: routeBodyMetadata.options.nullable,
+      required: routeBodyMetadata.options.required,
+    };
+
+    if (routeBodyMetadata.options.isArray === true) {
+      swaggerSchema = {
+        type: "array",
+        items: {
+          // eslint-disable-next-line @typescript-eslint/no-base-to-string
+          $ref: `#/components/schemas/${routeBodyMetadata.type}`,
+        },
+        deprecated: routeBodyMetadata.options.deprecated,
+        nullable: routeBodyMetadata.options.nullable,
+        required: routeBodyMetadata.options.required,
+      };
+    }
+  } else if (typeof routeBodyMetadata.type === typeof class {}) {
+    swaggerSchema = {
+      $ref: `#/components/schemas/${
+        (routeBodyMetadata.type as ClassType).name
+      }`,
+      deprecated: routeBodyMetadata.options.deprecated,
+      nullable: routeBodyMetadata.options.nullable,
+      required: routeBodyMetadata.options.required,
+    };
+    if (routeBodyMetadata.options.isArray === true) {
+      swaggerSchema = {
+        type: "array",
+        deprecated: routeBodyMetadata.options.deprecated,
+        nullable: routeBodyMetadata.options.nullable,
+        required: routeBodyMetadata.options.required,
+        items: {
+          $ref: `#/components/schemas/${
+            (routeBodyMetadata.type as ClassType).name
+          }`,
+        },
+      };
+    }
+    getReferenceSchema.call(this, routeBodyMetadata.type);
+  } else {
+    swaggerSchema = undefined;
+  }
+
+  return swaggerSchema;
+}
+/* eslint-enable no-invalid-this */
